foldhash::quality
Struct SeedableRandomState

pub struct SeedableRandomState { /* private fields */ }

A BuildHasher for quality::FoldHasher that is randomly initialized by default, but can also be initialized with a specific seed.

This can be useful for e.g. testing, but the downside is that this type has a size of 16 bytes rather than the 8 bytes RandomState is.

Implementations

impl SeedableRandomState

pub fn random() -> Self

Generates a random SeedableRandomState, similar to RandomState.

pub fn fixed() -> Self

Generates a fixed SeedableRandomState, similar to FixedState.

pub fn with_seed(per_hasher_seed: u64, shared_seed: &'static SharedSeed) -> Self

Generates a SeedableRandomState with the given per-hasher seed and SharedSeed.