foldhash

Struct SharedSeed

pub struct SharedSeed { /* private fields */ }

A random seed intended to be shared by many different foldhash instances.

This seed is consumed by FoldHasher::with_seed, and SeedableRandomState::with_seed.

Implementations

impl SharedSeed

pub fn global_random() -> &'static SharedSeed

Returns the globally shared randomly initialized SharedSeed as used by RandomState.

pub const fn global_fixed() -> &'static SharedSeed

Returns the globally shared fixed SharedSeed as used by FixedState.

pub const fn from_u64(seed: u64) -> Self

Generates a new SharedSeed from a single 64-bit seed.

Note that this is somewhat expensive so it is suggested to re-use the SharedSeed as much as possible, using the per-hasher seed to differentiate between hash instances.